requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.4
django-bulk-update==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
numpy==1.20.1
pandas==1.2.3
psycopg2==2.7.7
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
xlrd==1.2.0
